Add location filter to data processing
A new command line argument `--location-filter` was added to allow users to filter data by location coordinates (latitude, longitude and radius). This gives the user the ability to only process and forward data relevant to a specific geographical location. The README file was also updated to reflect this new feature.